+++ pyfixest/report/utils.py
@@ -1,6 +1,17 @@
 import re
 import warnings
-from typing import Optional
+from collections import Counter
+from collections.abc import ValuesView
+from typing import Optional, Union
+
+from pyfixest.estimation.feiv_ import Feiv
+from pyfixest.estimation.feols_ import Feols
+from pyfixest.estimation.fepois_ import Fepois
+from pyfixest.estimation.FixestMulti_ import FixestMulti
+
+ModelInputType = Union[
+    FixestMulti, Feols, Fepois, Feiv, list[Union[Feols, Fepois, Feiv]]
+]
 
 
 def _check_label_keys_in_covars(label_keys: list[str], covariate_names: list[str]):
@@ -167,3 +178,86 @@ def rename_event_study_coefs(coef_names_list: list):
         }
     """
     return {col: _rename_event_study_coefs(col) for col in coef_names_list}
+
+
+def _post_processing_input_checks(
+    models: ModelInputType,
+    check_duplicate_model_names: bool = False,
+    rename_models: Optional[dict[str, str]] = None,
+) -> list[Union[Feols, Fepois, Feiv]]:
+    """
+    Perform input checks for post-processing models.
+
+    Parameters
+    ----------
+        models : Union[List[Union[Feols, Fepois, Feiv]], FixestMulti]
+                The models to be checked. This can either be a list of models
+                (Feols, Fepois, Feiv) or a single FixestMulti object.
+        check_duplicate_model_names : bool, optional
+                Whether to check for duplicate model names. Default is False.
+                Mostly used to avoid overlapping models in plots created via
+                pf.coefplot() and pf.iplot().
+        rename_models : dict, optional
+                A dictionary to rename the models. The keys are the original model names
+                and the values are the new model names.
+
+    Returns
+    -------
+        List[Union[Feols, Fepois]]
+            A list of checked and validated models. The returned list contains only
+            Feols and Fepois types.
+
+    Raises
+    ------
+        TypeError: If the models argument is not of the expected type.
+
+    """
+    models_list: list[Union[Feols, Fepois, Feiv]] = []
+
+    if isinstance(models, (Feols, Fepois, Feiv)):
+        models_list = [models]
+    elif isinstance(models, FixestMulti):
+        models_list = models.to_list()
+    elif isinstance(models, (list, ValuesView)):
+        if all(isinstance(m, (Feols, Fepois, Feiv)) for m in models):
+            models_list = models
+        else:
+            raise TypeError(
+                "All elements in the models list must be instances of Feols, Feiv, or Fepois."
+            )
+    else:
+        raise TypeError("Invalid type for models argument.")
+
+    if check_duplicate_model_names or rename_models is not None:
+        all_model_names = [model._model_name for model in models_list]
+
+    if check_duplicate_model_names:
+        # create model_name_plot attribute to differentiate between models with the
+        # same model_name / model formula
+        for model in models_list:
+            model._model_name_plot = model._model_name
+
+        counter = Counter(all_model_names)
+        duplicate_model_names = [item for item, count in counter.items() if count > 1]
+
+        for duplicate_model in duplicate_model_names:
+            duplicates = [
+                model for model in models_list if model._model_name == duplicate_model
+            ]
+            for i, model in enumerate(duplicates):
+                model._model_name_plot = f"Model {i}: {model._model_name}"
+                warnings.warn(
+                    f"The _model_name attribute {model._model_name}' is duplicated for models in the `models` you provided. To avoid overlapping model names / plots, the _model_name_plot attribute has been changed to '{model._model_name_plot}'."
+                )
+
+        if rename_models is not None:
+            model_name_diff = set(rename_models.keys()) - set(all_model_names)
+            if model_name_diff:
+                warnings.warn(
+                    f"""
+                    The following model names specified in rename_models are not found in the models:
+                    {model_name_diff}
+                    """
+                )
+
+    return models_list
